requirements.txt
# Requirements #
django
djangorestframework
